Fixed responsive on Calendar and Shift views
