+++ src/scss/theme.scss
@@ -1,0 +1,8 @@
+$sidebar-background: #1e1e2e;
+$sidebar-selected-item: #5d5a88;
+
+$background-dart: #303446;
+
+$primary-color: #018241;
+$text-color: #666666;
+$text-color-dark: #000;
