+++ src/views/LoginView/LoginView.scss
@@ -5,7 +5,7 @@ body {
 }
 
 .login-background {
-  background-image: url('/src/assets/images/login-screen-background.png');
+  background-image: '@assets/images/login-screen-background.png';
   background-size: 100% 100%;
   background-repeat: no-repeat;
   height: 100%;
