+++ vite.config.js
@@ -12,6 +12,9 @@ export default defineConfig({
       '@components': path.resolve((__dirname, './src/components/')),
       '@views': path.resolve((__dirname, './src/views/')),
       '@mocks': path.resolve((__dirname, './mocks/')),
+      '@styles': path.resolve((__dirname, './src/scss')),
+      '@stores': path.resolve((__dirname, './src/stores')),
+      '@routes': path.resolve((__dirname, './src/routes'))
     },
   },
   test: {
